requirements.txt
astroid==2.15.5
black==23.3.0
certifi==2020.12.5
click==8.1.3
crccheck==1.0
dill==0.3.6
influxdb-client==1.16.0
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.0
pylint==2.17.4
python-dateutil==2.8.1
python-dotenv==1.0.0
pytz==2021.1
Rx==3.1.1
six==1.15.0
tomlkit==0.11.8
urllib3==1.26.5
wrapt==1.15.0
